Refactors Eleventy config to ESM and updates API spec handling
Migrates the Eleventy configuration file to ESM (`eleventy.config.mjs`).
This change allows for modern JavaScript syntax and module imports.

Updates API spec collection to correctly load and process YAML files.
This resolves issues with spec loading and ensures the specs are correctly built.

Moves Swagger UI and assets to a `public` folder and updates passthrough copy accordingly.

Updates the URL in the Swagger UI script to reflect the new location of the specifications.

Corrects the path to the annex document in the Dutch technical documentation.